+++ mslicer/src/windows/models/model.rs
@@ -6,136 +6,38 @@ use common::{
 };
 use const_format::concatcp;
 use egui::{
-    CollapsingHeader, Color32, Context, DragAndDrop, DragValue, Frame, Id, Label, LayerId, Order,
-    Popup, Response, ScrollArea, Sense, Stroke, TopBottomPanel, Ui, UiBuilder, Vec2, Widget,
-    emath::TSTransform, vec2,
+    CollapsingHeader, Color32, Context, DragValue, Label, Popup, Response, Sense, Ui, UiBuilder,
+    Widget, vec2,
 };
 use egui_phosphor::regular::{
-    ARROW_LINE_DOWN, COPY, CURSOR_TEXT, DICE_THREE, EYE, EYE_SLASH, LINK_BREAK, LINK_SIMPLE, TRASH,
-    WARNING,
+    ARROW_LINE_DOWN, COPY, CURSOR_TEXT, DICE_THREE, EYE, EYE_SLASH, FOLDER_DASHED, LINK_BREAK,
+    LINK_SIMPLE, TRASH, WARNING,
 };
 use nalgebra::Vector3;
 
 use crate::{
     app::{App, history::ModelAction},
-    project::model::{MeshWarnings, RenameState},
+    project::{
+        Collection,
+        model::{MeshWarnings, RenameState},
+    },
     ui::components::{
         being_edited, grid, history_tracked_model, vec3_dragger, vec3_dragger_proportional,
     },
+    windows::models::Action,
 };
 
 const WARN_NON_MANIFOLD: &str = "This mesh is non-manifold, it may produce unexpected results when sliced.\nConsider running it through a mesh repair tool.";
 const WARN_OUT_OF_BOUNDS: &str = "This mesh extends beyond the printer volume and will be cut off.";
 
-enum Action {
-    None,
-    Remove(usize),
-    Duplicate(usize),
-}
-
-struct DraggedModel {
-    index: usize,
-    offset: Vec2,
-}
-
-pub fn ui(app: &mut App, ui: &mut Ui, ctx: &Context) {
-    if app.project.models.is_empty() {
-        ui.vertical_centered(|ui| {
-            ui.label("No models loaded yet.");
-        });
-        return;
-    }
-
-    if let Some(id) = app.state.selected_model
-        && let Some(model) = app.project.models.iter_mut().position(|x| x.id == id)
-    {
-        let mut action = Action::None;
-        TopBottomPanel::bottom("model_props")
-            .resizable(false)
-            .frame(Frame::new().inner_margin(2.0))
-            .default_height(f32::MAX)
-            .show_inside(ui, |ui| {
-                ui.add_space(4.0);
-                model_properties(app, ui, ctx, &mut action, model)
-            });
-
-        match action {
-            Action::Remove(i) => {
-                app.project.models.remove(i);
-            }
-            Action::Duplicate(i) => {
-                let model = app.project.models[i].clone();
-                app.project.models.push(model);
-            }
-            Action::None => {}
-        }
-    }
-
-    ScrollArea::vertical().show(ui, |ui| {
-        for i in 0..app.project.models.len() {
-            let id = app.project.models[i].id;
-            let eid = Id::new("model").with(id);
-
-            if ctx.is_being_dragged(eid) {
-                let layer_id = LayerId::new(Order::Tooltip, eid);
-                let response = ui.scope_builder(UiBuilder::new().layer_id(layer_id), |ui| {
-                    model_entry(app, ui, i, true)
-                });
-
-                if let Some(pointer_pos) = ctx.pointer_interact_pos()
-                    && let Some(dragged) = DragAndDrop::payload::<DraggedModel>(ctx)
-                {
-                    let delta = pointer_pos - response.response.rect.center() - dragged.offset;
-                    ctx.transform_layer_shapes(layer_id, TSTransform::from_translation(delta));
-                }
-            } else {
-                let response = model_entry(app, ui, i, false);
-                if response.drag_started() {
-                    ctx.set_dragged_id(eid);
-                    let offset = (ctx.pointer_interact_pos())
-                        .map(|p| p - response.rect.center())
-                        .unwrap_or_default();
-                    let payload = DraggedModel { index: i, offset };
-
-                    DragAndDrop::set_payload(ctx, payload);
-                }
-            }
-        }
-    });
-
-    if let Some(pointer) = ctx.pointer_interact_pos()
-        && let Some(dragged) = DragAndDrop::payload::<DraggedModel>(ctx)
-    {
-        let rect = ui.max_rect();
-
-        let stroke = Stroke::new(1.0, Color32::WHITE);
-        let line = |y| ui.painter().hline(rect.x_range(), y, stroke);
-
-        let entry_height = 18.0 + ui.style().spacing.item_spacing.y;
-        let t = (pointer.y - rect.min.y) / entry_height + 0.5;
-        let new_index = (t as usize).min(app.project.models.len());
-
-        line(rect.min.y + new_index as f32 * entry_height);
-
-        if ctx.input(|i| i.pointer.any_released()) {
-            let insert_index = new_index - (dragged.index < new_index) as usize;
-            let model = app.project.models.remove(dragged.index);
-            app.project.models.insert(insert_index, model);
-        }
-    }
-
-    let (_rect, response) = ui.allocate_exact_size(ui.available_size(), Sense::click());
-    response.clicked().then(|| app.state.selected_model = None);
-}
-
-fn model_entry(app: &mut App, ui: &mut Ui, idx: usize, dragged: bool) -> Response {
+pub fn model_entry(app: &mut App, ui: &mut Ui, idx: usize, dragged: bool) -> Response {
     let model = &mut app.project.models[idx];
     let id = model.id;
 
     let (rect, response) =
         ui.allocate_exact_size(vec2(ui.available_width(), 18.0), Sense::click_and_drag());
 
-    let selected = Some(id) == app.state.selected_model;
+    let selected = app.state.selected.contains_model(id);
     let color = if selected && !dragged {
         ui.visuals().selection.bg_fill
     } else if response.hovered() || dragged {
@@ -150,14 +52,15 @@ fn model_entry(app: &mut App, ui: &mut Ui, idx: usize, dragged: bool) -> Respons
     ui.painter().rect_filled(rect_margin, 2.0, color);
 
     if response.clicked() {
-        if selected {
-            app.state.selected_model = None;
-        } else {
-            app.state.selected_model = Some(id);
-        }
+        app.state
+            .selected
+            .model_clicked(id, ui.input(|x| x.modifiers.shift));
     }
 
-    ui.scope_builder(UiBuilder::new().max_rect(rect), |ui| {
+    let indent = !dragged && model.collection.is_some();
+    let ui_builder =
+        UiBuilder::new().max_rect(rect.with_min_x(rect.min.x + [0.0, 12.0][indent as usize]));
+    ui.scope_builder(ui_builder, |ui| {
         ui.horizontal(|ui| {
             ui.visuals_mut().override_text_color = selected.then_some(Color32::WHITE);
             ui.visuals_mut().button_frame = false;
@@ -210,7 +113,7 @@ fn model_entry(app: &mut App, ui: &mut Ui, idx: usize, dragged: bool) -> Respons
     response
 }
 
-fn model_properties(app: &mut App, ui: &mut Ui, ctx: &Context, action: &mut Action, i: usize) {
+pub fn model_properties(app: &mut App, ui: &mut Ui, ctx: &Context, action: &mut Action, i: usize) {
     let model = &mut app.project.models[i];
 
     let platform = &app.project.slice_config.platform_size;
@@ -222,6 +125,13 @@ fn model_properties(app: &mut App, ui: &mut Ui, ctx: &Context, action: &mut Acti
         (ui.button(concatcp!(TRASH, " Delete")).clicked()).then(|| *action = Action::Remove(i));
         (ui.button(concatcp!(COPY, " Duplicate")).clicked())
             .then(|| *action = Action::Duplicate(i));
+        ui.button(concatcp!(FOLDER_DASHED, " Collect"))
+            .clicked()
+            .then(|| {
+                let collection = Collection::new("Collection".into());
+                model.collection = Some(collection.id);
+                app.project.collections.push(collection);
+            });
         ui.button(concatcp!(ARROW_LINE_DOWN, " Align to Bed"))
             .clicked()
             .then(|| {